refactor(suite): rename stake.ts to ethereumStaking.ts for clarity
