+++ packages/suite/src/components/suite/StakingProcess/StakingInfo.tsx
@@ -13,7 +13,7 @@ import {
 } from '@suite-common/wallet-core';
 
 import { Translation } from 'src/components/suite';
-import { getDaysToAddToPool } from 'src/utils/suite/stake';
+import { getDaysToAddToPool } from 'src/utils/suite/ethereumStaking';
 import { CoinjoinRootState } from 'src/reducers/wallet/coinjoinReducer';
 
 import { InfoRow } from './InfoRow';
