+++ packages/suite/src/components/suite/modals/ReduxModal/UserContextModal/StakeModal/StakeEthForm/Inputs.tsx
@@ -15,7 +15,7 @@ import {
     validateReserveOrBalance,
 } from 'src/utils/suite/validation';
 import { FIAT_INPUT, CRYPTO_INPUT } from 'src/types/wallet/stakeForms';
-import { validateStakingMax } from 'src/utils/suite/stake';
+import { validateStakingMax } from 'src/utils/suite/ethereumStaking';
 import { FormFractionButtons } from 'src/components/suite/FormFractionButtons';
 
 export const Inputs = () => {
