+++ packages/suite/src/utils/suite/__fixtures__/ethereumStaking.ts
